feat(react): 优化react适配逻辑
